+++ Pandora/Paradigm/Structure.hs
@@ -5,5 +5,5 @@ import Pandora.Paradigm.Structure.Stack as Exports
 import Pandora.Paradigm.Basis.Cofree (Cofree)
 import Pandora.Paradigm.Basis.Junction.Transformer (type (:>:))
 
-type family Nonempty structure :: * where
-	Nonempty ((Cofree :>: t) a) = Cofree t a
+type family Nonempty structure a :: * where
+	Nonempty (Cofree :>: t) a = Cofree t a