+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 * Define `attached` destructor for `Product` datatype
 * Define `empty` destructor for `Stack` structure
 * Replace `flip` on `?` type operator
+* Remove type parameter from `Stack` type synonymous
+* Split `structure` and `value` in `Nonempty` type family